Strip down News menu item
It felt kind of clumsy having the 'News' menu item being a dropdown with
the two elements of 'Overview' and 'Archive'.

Now, the 'News' menu item is a simple link pointing to the news index
page. The news archive is accessible either via the news index page (at
the very bottom) or the landing page.

Signed-off-by: Torbjörn Klatt <t.klatt@fz-juelich.de>