Make collapsable event/project meta more awesome
Instead of texts 'hide'/'show profile' there are now little icons
'-'/'i'. Also, the icon to show the profile is now in the
sub-navigation, which vertically lines up with the aside meta's heading.

I dropped the usage of TWBS' `.btn` classes on these icons, as they were
causing significant delays in rendering the transition.

Signed-off-by: Torbjörn Klatt <t.klatt@fz-juelich.de>