No sub-pages for projects
This will either mess up the report or lead to confusion on whether the
sub-pages will be part of the report or not.

As I nuked the sub-navigation bar for project pages, the meta aside is
not collapsible any more as I have no idea where to place the 'Show
Profile' button.

Signed-off-by: Torbjörn Klatt <t.klatt@fz-juelich.de>